requirements.txt
fastapi
uvicorn
transformers
torch
flash-attn; platform_system == "Linux"
accelerate
diffusers @ git+https://github.com/huggingface/diffusers
sentencepiece
httpx
pillow
pybase64
sglang==0.5.5
modelscope
orjson
uvloop; sys_platform != "win32"
psutil
pyzmq
safetensors
